+++ docs/HISTORY.rst
@@ -10,6 +10,8 @@ Changelog
 
 - Only export point coordinates. Fixes #81.
 
+- Show import source in event view. Implements #78.
+
 1.3.2 (2014-09-18)
 ~~~~~~~~~~~~~~~~~~